Split key_io (address/key encodings) off from base58
