+++ src/bitcoin-tx.cpp
@@ -6,11 +6,11 @@
 #include <config/bitcoin-config.h>
 #endif
 
-#include <base58.h>
 #include <clientversion.h>
 #include <coins.h>
 #include <consensus/consensus.h>
 #include <core_io.h>
+#include <key_io.h>
 #include <keystore.h>
 #include <policy/policy.h>
 #include <policy/rbf.h>
