+++ src/core_write.cpp
@@ -4,9 +4,9 @@
 
 #include <core_io.h>
 
-#include <base58.h>
 #include <consensus/consensus.h>
 #include <consensus/validation.h>
+#include <key_io.h>
 #include <script/script.h>
 #include <script/standard.h>
 #include <serialize.h>
