+++ src/httprpc.cpp
@@ -4,9 +4,9 @@
 
 #include <httprpc.h>
 
-#include <base58.h>
 #include <chainparams.h>
 #include <httpserver.h>
+#include <key_io.h>
 #include <rpc/protocol.h>
 #include <rpc/server.h>
 #include <random.h>
