+++ src/qt/bitcoinaddressvalidator.cpp
@@ -4,7 +4,7 @@
 
 #include <qt/bitcoinaddressvalidator.h>
 
-#include <base58.h>
+#include <key_io.h>
 
 /* Base58 characters are:
      "123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz"
