+++ src/qt/coincontroldialog.cpp
@@ -15,6 +15,7 @@
 
 #include <wallet/coincontrol.h>
 #include <init.h>
+#include <key_io.h>
 #include <policy/fees.h>
 #include <policy/policy.h>
 #include <validation.h> // For mempool
