+++ src/qt/paymentrequestplus.cpp
@@ -9,6 +9,7 @@
 
 #include <qt/paymentrequestplus.h>
 
+#include <script/script.h>
 #include <util.h>
 
 #include <stdexcept>
