+++ src/qt/paymentrequestplus.h
@@ -10,7 +10,8 @@
 #include <qt/paymentrequest.pb.h>
 #pragma GCC diagnostic pop
 
-#include <base58.h>
+#include <amount.h>
+#include <script/script.h>
 
 #include <openssl/x509.h>
