+++ src/qt/transactionrecord.cpp
@@ -4,8 +4,8 @@
 
 #include <qt/transactionrecord.h>
 
-#include <base58.h>
 #include <consensus/consensus.h>
+#include <key_io.h>
 #include <validation.h>
 #include <timedata.h>
 #include <wallet/wallet.h>
