+++ src/qt/walletmodel.cpp
@@ -14,8 +14,8 @@
 #include <qt/sendcoinsdialog.h>
 #include <qt/transactiontablemodel.h>
 
-#include <base58.h>
 #include <chain.h>
+#include <key_io.h>
 #include <keystore.h>
 #include <validation.h>
 #include <net.h> // for g_connman
