+++ src/rpc/rawtransaction.cpp
@@ -3,7 +3,6 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <base58.h>
 #include <chain.h>
 #include <coins.h>
 #include <consensus/validation.h>
@@ -12,6 +11,7 @@
 #include <keystore.h>
 #include <validation.h>
 #include <validationinterface.h>
+#include <key_io.h>
 #include <merkleblock.h>
 #include <net.h>
 #include <policy/policy.h>
