+++ src/rpc/server.cpp
@@ -5,9 +5,9 @@
 
 #include <rpc/server.h>
 
-#include <base58.h>
 #include <fs.h>
 #include <init.h>
+#include <key_io.h>
 #include <random.h>
 #include <sync.h>
 #include <ui_interface.h>
