+++ src/wallet/rpcdump.cpp
@@ -2,8 +2,8 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <base58.h>
 #include <chain.h>
+#include <key_io.h>
 #include <rpc/safemode.h>
 #include <rpc/server.h>
 #include <wallet/init.h>
