+++ src/wallet/walletdb.cpp
@@ -5,10 +5,10 @@
 
 #include <wallet/walletdb.h>
 
-#include <base58.h>
 #include <consensus/tx_verify.h>
 #include <consensus/validation.h>
 #include <fs.h>
+#include <key_io.h>
 #include <protocol.h>
 #include <serialize.h>
 #include <sync.h>
